requirements.txt
anyio==3.6.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
autopep8==2.0.1
Babel==2.8.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
comm==0.1.2
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.6.5
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
Django==4.1.5
django-cleanup==6.0.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.1.0
docopt==0.6.2
drf-standardized-errors==0.12.4
duplicity==0.8.21
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.2.0
fake-useragent==1.1.1
fastapi==0.89.1
fasteners==0.14.1
fastjsonschema==2.16.2
ffmpy==0.3.0
filelock==3.6.0
fqdn==1.5.1
future==0.18.2
h11==0.14.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
jeepney==0.7.1
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libvirt-python==8.0.0
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.8
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.2
notebook_shim==0.2.2
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
packaging==23.0
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.12.4
proton-client==0.7.1
protonvpn-cli==2.2.11
protonvpn-gui==1.12.0
protonvpn-nm-lib==3.14.0
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycodestyle==2.10.0
pycparser==2.21
pycups==2.0.1
pydantic==1.10.4
pydub==0.25.1
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyotp==2.8.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
PySocks==1.7.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.21.0
python-gnupg==0.4.8
python-json-logger==2.0.4
pythondialog==3.5.3
PythonTurtle==0.3.2
pytz==2022.7
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
recaptcha-solver==1.0.4
reportlab==3.6.8
requests==2.28.1
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
SecretStorage==3.3.1
selenium==4.7.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.9.0
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.22.0
stem==1.8.1
systemd-python==234
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
undetected-chromedriver==3.2.1
uri-template==1.2.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wcwidth==0.2.6
webcolors==1.12
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.4.2
websockets==10.4
widgetsnbextension==4.0.5
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
